+++ src/memory/alloc.c
@@ -29,10 +29,10 @@ void MemoryAllocated(void *addr, size_t size, const char function[200], uint32_t
 
 restart:
   for(mti = &MemoryTableIndices; mti != NULL; mti = mti->next) {
-    for(mt = mti->addr; (uint32_t) mt < (uint32_t) mti->addr + mti->size; mt = (MemoryTable_t *) ((uint32_t) mt + sizeof(MemoryTable_t))) {
+    for(mt = mti->addr; (uint64_t) mt < (uint64_t) mti->addr + mti->size; mt = (MemoryTable_t *) ((uint64_t) mt + sizeof(MemoryTable_t))) {
       if(mt->status != 1) {
         // Not allocated
-        mt->addr = (uint32_t) addr;
+        mt->addr = (uint64_t) addr;
         mt->size = size;
         mt->status = 1;
         strcpy(mt->function, function);
@@ -42,7 +42,7 @@ restart:
 
       // If the iteration covers one of the last 5 entries in the final table, then create a new table
       // The process of creating a new table requires the use of malloc()
-      if(mti->next == NULL && (uint32_t) mt >= (uint32_t) mti->addr + mti->size - (sizeof(MemoryTable_t) * 5) && new_table_pending == 0) {
+      if(mti->next == NULL && (uint64_t) mt >= (uint64_t) mti->addr + mti->size - (sizeof(MemoryTable_t) * 5) && new_table_pending == 0) {
         new_table_pending = 1;
         MemoryNewTable();
         new_table_pending = 0;
